requirements.txt
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.1.2
Flask==3.0.0
werkzeug==3.0.1
tensorflow==2.13.0
Pillow==10.1.0
Flask-HTTPAuth==4.8.0
sklearn==1.3.2